requirements_unfrozen.txt
python-dotenv

streamlit
streamlit-extras
streamlit-oauth==0.1.9

langchain
langchain_core
langchain-community
langchain-openai
langchain_anthropic
openai